requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-swagger-ui==4.11.1
apscheduler==3.10.4
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
webdriver-manager==4.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # PostgreSQL driver
# tkinter is usually included with Python installation